requirements.txt
asgiref==3.8.1
Django==5.1.3
django-ckeditor-5==0.2.15
django-daisy==0.1.28
djangorestframework==3.15.2
pillow==11.0.0
sqlparse==0.5.2
typing_extensions==4.12.2
